// AAInsights-UI/src/models/ConsentDetails.ts
export class ConsentDetails{

    position:number;
    id:string;
    fiu:string;
    date:Date;
    purpose:string;
    consentType:string;
    // dateRange: string;
    consentMode: string;
    
    // operator: string;
    // dataFilterType: string;
    // dataFilterValue: string;
    dataLifeUnit: string;
    dataLifeValue: string;
    fiDataFrom: string;
    fiDataTo: string;
    frequencyUnit: string;
    frequencyValue: number;
    purposeText: string;
    purposeCode: number;
    consentStart: string;
    consentExpiry: string;
    // 'consentMode': string;
    fetchType: string;
    // 'consentType': string;
    fiType: string;
    sensitivity:string;
    alert:string;
    fiDateMonths:number;
    action:string;
}